# Cache the cache to speed up ethash