```
BEFORE:
In [1]: from ethereum.pow.ethpow import get_cache

In [2]: %timeit get_cache(1,1)
624 ms ± 15.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

AFTER:
In [2]: from ethereum.pow.ethpow import get_cache

In [3]: %timeit get_cache(1,1)
The slowest run took 4.25 times longer than the fastest. This could mean that an intermediate result is being cached.
6.82 µs ± 4.68 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [4]: %%timeit
   ...: try:
   ...:     validate_seal(RootBlockHeader(), ConsensusType.POW_ETHASH)
   ...: except Exception:
   ...:     pass
   ...: 
2.63 ms ± 326 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```